:bug: Fix filter documentation generation
Choicefields did not get their enums if they were explicitly declared
on the filterset, and thereby assumed to not map to a model field.

The model field is now attempted to retrieve, and the exception
caught if it's missing. Code is made conditional on there being
a model field, otherwise falling back to different defaults.