:recycle: Overhaul of notifications and atomicity
Wrapping the viewset actions create/update/destroy in an atomic block is
_not_ always desired, as this blocks manual transaction management in
subclasses implementing this mixin (unless they go overriding the
methods themselves.

A practical case for this is fine-grained transaction management in the
ZaakInformatieObject serializer create method, where the remote
OIO is created - so the ZIO itself must be visible to another
request-response cycle. This is currently not possible.

This patch keeps the default behaviour by wrapping everything in
an atomic transaction, but it gives the flexibility to subclasses
via the `notifications_wrap_in_atomic_block` attribute to control
this behaviour and deal with transactions themselves.

Context for this fix: https://github.com/open-zaak/open-zaak/issues/819