:bug: Can't recalculate and save ETags on deleted objects
Objects may be scheduled by the dependency tracker for update, but at
the same time have been deleted in the transaction because of a
cascading delete. This leads to database errors because the record
no longer exists and can't be updated.

Our transaction.on_commit callback now checks if the object still
exists or not before doing the recalculation.