:bug: Resolve migration merge conflicts
Upgrading from the 1.0.x series to 1.7.x is pretty much broken because
of migration differences. Both branches contain migrations that do the
same physical changes in the database (drop request_id, add logrecord_id
columns). The migration files from 1.0.x were not included in the
1.7.x series, or vice versa.

This PR adds the migration files from 1.0.x into the latest release
series, and a merge migration file to reconcile the migration tree.
This causes problems for downstream projects upgrading from 1.0.x to
latest, but also for projects that have never been on 1.0.x, as the
migrations contain conflicting operations. We cannot simply update the
code of one of the two migration branches, since there are projects
that have performed the request_id/logrecord_id migrations, but also
projects that haven't and we cannot know upfront which has what.

To solve this, custom AddField and RemoveField migration operations
are defined and they replace the default operations in the migration
files. Their implementation uses the schema inspector for the database
used to figure out if the relevant columns still/already exist and
adapt their actual database changes based on that introspection.