🔒 Fix JWT verification
A JWT has a header field `alg` that specifies the algorithm used in the
signature.

PyJWT checks this with `alg in algorithms`, this "works" because
`"HS256" in "HS256"` is true, but so is `"" in "HS256"` and
`"HS2" in "HS256"`.

Luckily there currently are no PyJWT algorithms like that.
There is no HMAC SHA2, and the Null encryption is named "none" not "".